Create Basic Home UI (#8)
I created a preliminary frontend for the homepage (i.e. files, pastes,
URLs). There are definitely some improvements that we will need to make,
but I think this is a pretty decent starting point for the UI. For now,
I routed the page to `/home`, but that can easily be changed.