Initial implementation of paste list (#11)
Kept this one really simple since we don't know exactly what we want to
display with each paste. I decided to make a table component(with a head
and body) in case it could be of use somewhere else in the project. The
table allows for alignment and width distribution depending on however
the user wants his attributes to appear.

Added todo correlating to an issue in github for a view paste page or
modal of some sort.

I think we also have to talk about the organization of the repo, where
new pages and components are going. Not sure if this will have merge
conflicts based on what John did earlier today