User/johnflanagan827/shorten url frontend (#31)
I've implemented a basic version of the URL shortening feature on the
frontend (closes #19). Currently, when the "shorten" button is clicked,
it simply displays the shortened code. The design of this feature,
especially how the code is presented, will need improvement later on. As
of now, there is no validation for the URLs entered, which is something
we will need to address. Also, we still need to create the option to use
customize URL. However, my main focus was to prepare the core
functionality of generating short codes in time for our demo.

Also, here is an example of a generated shortcode on the frontend:


![image](https://github.com/DocuDump/DocuDump/assets/69359897/bade18df-bcd0-4e80-a6d0-da7e879f0664)