Custom URL support (#38)
Added preliminary support for custom URLs (closes #33). I added
preliminary error handling for if the custom URL already exists (see
picture below). For now, it just does the main functionality of
displaying the custom URL code if it was successful, and giving a red
error message if it was not successful.

We still eventually need to add error handling for invalid URLs (#32)
and create better UI for when the short code / custom URL is generated
(#34).

- Here is an example of an invalid custom URL:


![image](https://github.com/DocuDump/DocuDump/assets/69359897/682066cf-b717-443a-af5e-a1bc8a445a4f)