Completed issues 48 and 51 (#52)
Made a change to settings.tsx to disable the password on
public pages
I also noticed the form to create the paste was being submitted when the
user clicked the visibility button so I fixed that.

I added a redirect from the / to /home. This is subject to change if we
create a login or landing page of some sort.